refactor option types into params.options.types
