+++ aiida/cmdline/params/options/types/__init__.py
@@ -1,0 +1,3 @@
+from conditional import ConditionalOption
+from interactive import InteractiveOption
+from multivalue import MultipleValueOption
